feat: change the project folder hierarchy
Also, export types from the app and make the custom hook a default export
